+++ app/index.js
@@ -16,15 +16,15 @@ var richMessage = require('rich-message')
 var Swarm = require('friends-swarm')
 var yo = require('yo-yo')
 
-var config = require('./config')
-var util = require('./lib/util')
-var command = require('./lib/command')
-var Signature = require('./lib/signature')
-var Channels = require('./lib/elements/channels')
-var Composer = require('./lib/elements/composer')
-var Messages = require('./lib/elements/messages')
-var Status = require('./lib/elements/status')
-var Users = require('./lib/elements/users')
+var config = require('../config')
+var util = require('./util')
+var command = require('./command')
+var Signature = require('./signature')
+var Channels = require('./elements/channels')
+var Composer = require('./elements/composer')
+var Messages = require('./elements/messages')
+var Status = require('./elements/status')
+var Users = require('./elements/users')
 
 inherits(App, EventEmitter)
