+++ app/style/index.styl
@@ -98,7 +98,7 @@ button
   width: SIDEBAR_WIDTH
   height: 100%
   color: GRAY_LIGHTEST
-  background: BLUE_DARK top left url('../static/logo.png') no-repeat
+  background: BLUE_DARK top left url('../../static/logo.png') no-repeat
   background-size: SIDEBAR_WIDTH
   padding-top: FRIENDS_LOGO_HEIGHT
   padding-bottom: SIDEBAR_BOTTOM_HEIGHT
@@ -247,7 +247,7 @@ button
     color: inherit
   .verified
     display: inline-block
-    background-image: url('../static/verified.svg')
+    background-image: url('../../static/verified.svg')
     background-size: 10px
     width: 10px
     height: 10px
