+++ app/web.js
@@ -1,7 +1,7 @@
 module.exports = Web
 
 var inherits = require('util').inherits
-var App = require('./app.js')
+var App = require('./')
 
 inherits(Web, App)
