+++ app/windows/git-help.js
